+++ internal/dto/sendcoin_dto.go
@@ -2,5 +2,5 @@ package dto
 
 type SendCoinRequest struct {
 	ToUser string `json:"toUser"`
-	Amount int    `json:"amount"`
+	Amount int64  `json:"amount"`
 }
